+++ packages/react-native-template-realm-ts/template/app/components/TaskItem.tsx
@@ -1,45 +1,39 @@
-import React, { memo } from 'react';
-import { View, Text, Pressable, Platform, StyleSheet } from 'react-native';
+import React, {memo} from 'react';
+import {View, Text, Pressable, Platform, StyleSheet} from 'react-native';
 
 import colors from '../styles/colors';
 
 interface TaskItemProps {
-  description: string,
-  isComplete: boolean,
-  onToggleStatus: () => void,
-  onDelete: () => void
+  description: string;
+  isComplete: boolean;
+  onToggleStatus: () => void;
+  onDelete: () => void;
 }
 
-function TaskItem({ description, isComplete, onToggleStatus, onDelete }: TaskItemProps) {
+function TaskItem({
+  description,
+  isComplete,
+  onToggleStatus,
+  onDelete,
+}: TaskItemProps) {
   return (
     <View style={styles.task}>
       <Pressable
         onPress={onToggleStatus}
-        style={[styles.status, isComplete && styles.completed]}
-      >
-        <Text style={styles.icon}>
-          {isComplete ? '✓' : '○'}
-        </Text>
+        style={[styles.status, isComplete && styles.completed]}>
+        <Text style={styles.icon}>{isComplete ? '✓' : '○'}</Text>
       </Pressable>
       <View style={styles.descriptionContainer}>
-        <Text
-          numberOfLines={1}
-          style={styles.description}
-        >
+        <Text numberOfLines={1} style={styles.description}>
           {description}
         </Text>
       </View>
-      <Pressable
-        onPress={onDelete}
-        style={styles.deleteButton}
-      >
-        <Text style={styles.deleteText}>
-          Delete
-        </Text>
+      <Pressable onPress={onDelete} style={styles.deleteButton}>
+        <Text style={styles.deleteText}>Delete</Text>
       </Pressable>
     </View>
   );
-};
+}
 
 const styles = StyleSheet.create({
   task: {
@@ -54,15 +48,15 @@ const styles = StyleSheet.create({
         shadowColor: colors.black,
         shadowOffset: {
           width: 0,
-          height: 4
+          height: 4,
         },
         shadowOpacity: 0.7,
-        shadowRadius: 3
+        shadowRadius: 3,
       },
       android: {
-        elevation: 3
-      }
-    })
+        elevation: 3,
+      },
+    }),
   },
   descriptionContainer: {
     flex: 1,
@@ -71,7 +65,7 @@ const styles = StyleSheet.create({
   description: {
     paddingHorizontal: 10,
     color: colors.black,
-    fontSize: 17
+    fontSize: 17,
   },
   status: {
     width: 50,
@@ -79,10 +73,10 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     borderTopLeftRadius: 5,
     borderBottomLeftRadius: 5,
-    backgroundColor: colors.gray
+    backgroundColor: colors.gray,
   },
   completed: {
-    backgroundColor: colors.purple
+    backgroundColor: colors.purple,
   },
   deleteButton: {
     justifyContent: 'center',
@@ -90,20 +84,22 @@ const styles = StyleSheet.create({
   deleteText: {
     marginHorizontal: 10,
     color: colors.gray,
-    fontSize: 17
+    fontSize: 17,
   },
   icon: {
     color: colors.white,
     textAlign: 'center',
     fontSize: 17,
-    fontWeight: 'bold'
-  }
+    fontWeight: 'bold',
+  },
 });
 
 // We want to make sure only tasks that change are rerendered
-const shouldNotRerender = (prevProps: TaskItemProps, nextProps: TaskItemProps) => (
-  prevProps.description === nextProps.description
-  && prevProps.isComplete === nextProps.isComplete
-);
+const shouldNotRerender = (
+  prevProps: TaskItemProps,
+  nextProps: TaskItemProps,
+) =>
+  prevProps.description === nextProps.description &&
+  prevProps.isComplete === nextProps.isComplete;
 
 export default memo(TaskItem, shouldNotRerender);